chore: make driver just an alias of GmailDriver
