+++ src/platform-implementation-js/dom-driver/inbox/views/inbox-drawer-view.d.ts
@@ -1,4 +1,13 @@
-import type { DrawerViewOptions } from '../../../driver-interfaces/driver';
+import ComposeView from '../../../views/compose-view';
+
+export interface DrawerViewOptions {
+  el: HTMLElement;
+  title?: string;
+  chrome?: boolean;
+  composeView?: ComposeView;
+  closeWithCompose?: boolean;
+  matchSidebarContentPanelWidth?: boolean;
+}
 
 export default class InboxDrawerView {
   constructor(options: DrawerViewOptions);
