+++ src/platform-implementation-js/driver-interfaces/driver.js
@@ -1,12 +1,7 @@
 /* @flow */
 
 import type Kefir from 'kefir';
-import type KeyboardShortcutHandle from '../views/keyboard-shortcut-handle';
-import type Logger from '../lib/logger';
-import type { ComposeViewDriver } from './compose-view-driver';
 import type ComposeView from '../views/compose-view';
-import type { RouteViewDriver } from './route-view-driver';
-import type CommonPageCommunicator from '../lib/common-page-communicator';
 
 // import type GmailBackdrop from '../dom-driver/gmail/views/gmail-backdrop';
 import type InboxBackdrop from '../dom-driver/inbox/views/inbox-backdrop';
@@ -14,13 +9,6 @@ export type Backdrop = Object | InboxBackdrop;
 
 import type GmailThreadView from '../dom-driver/gmail/views/gmail-thread-view';
 export type ThreadViewDriver = GmailThreadView;
-
-import type { ThreadRowViewDriver } from './thread-row-view-driver';
-
-import type { MoleViewDriver } from './mole-view-driver';
-
-import type { MessageViewDriver } from './message-view-driver';
-import type ContentPanelViewDriver from '../driver-common/sidebar/ContentPanelViewDriver';
 
 import type GmailAttachmentCardView from '../dom-driver/gmail/views/gmail-attachment-card-view';
 export type AttachmentCardViewDriver = GmailAttachmentCardView;
@@ -36,84 +24,10 @@ export type DrawerViewOptions = {
 import type InboxDrawerView from '../dom-driver/inbox/views/inbox-drawer-view';
 export type DrawerViewDriver = InboxDrawerView;
 
-import type { PiOpts } from '../platform-implementation';
-import type GmailSupportItemView, {
-  SupportItemDescriptor,
-} from '../dom-driver/gmail/views/gmail-support-item-view';
-import type GmailNavItemView from '../dom-driver/gmail/views/gmail-nav-item-view';
-import type GmailAppMenuItemView from '../dom-driver/gmail/views/gmail-app-menu-item-view';
+import type GmailDriver from '../dom-driver/gmail/gmail-driver';
 
 // TODO fill in some of these any types
-export type Driver = {
-  onready: Promise<void>,
-  getLogger(): Logger,
-  getAppId(): string,
-  getOpts(): PiOpts,
-  getRouteViewDriverStream(): Kefir.Observable<RouteViewDriver>, // should be a property
-  getRowListViewDriverStream(): Kefir.Observable<Object>,
-  openNewComposeViewDriver(): Promise<ComposeViewDriver>,
-  getNextComposeViewDriver(): Promise<ComposeViewDriver>,
-  getComposeViewDriverStream(): Kefir.Observable<ComposeViewDriver>,
-  openDraftByMessageID(messageID: string): Promise<void>,
-  getThreadViewDriverStream(): Kefir.Observable<ThreadViewDriver>,
-  getMessageViewDriverStream(): Kefir.Observable<MessageViewDriver>,
-  getAttachmentCardViewDriverStream(): Kefir.Observable<AttachmentCardViewDriver>,
-  activateShortcut(
-    keyboardShortcutHandle: KeyboardShortcutHandle,
-    appName: ?string,
-    appIconUrl: ?string
-  ): void,
-  getPageCommunicator(): CommonPageCommunicator,
-  getGmailActionToken(): Promise<string>,
-  getUserEmailAddress(): string,
-  isConversationViewDisabled(): Promise<boolean>,
-  getUserLanguage(): string,
-  getUserContact(): Contact,
-  getAccountSwitcherContactList(): Contact[],
-  getThreadRowViewDriverStream(): Kefir.Observable<ThreadRowViewDriver>,
-  registerThreadButton(options: Object): () => void,
-  addNavItem(
-    appId: string,
-    navItemDescriptor: Object
-  ): Promise<GmailNavItemView>,
-  addAppMenuItem(
-    appId: string,
-    menuItemDescriptor: Object
-  ): Promise<GmailAppMenuItemView>,
-  addSupportItem(
-    supportItemDescriptor: SupportItemDescriptor
-  ): GmailSupportItemView,
-  getSentMailNativeNavItem(): Promise<Object>,
-  createLink(a: any, b: any): any,
-  goto(routeID: string, params: any): Promise<void>,
-  addCustomRouteID(routeID: string): () => void,
-  addCustomListRouteID(routeID: string, handler: Function): () => void,
-  showCustomRouteView(el: HTMLElement): void,
-  setShowNativeNavMarker(isNative: boolean): any,
-  setShowNativeAddonSidebar(isNative: boolean): any,
-  registerSearchSuggestionsProvider(handler: Function): void,
-  registerSearchQueryRewriter(obj: Object): void,
-  addToolbarButtonForApp(
-    buttonDescriptor: Kefir.Observable<Object>
-  ): Promise<Object>,
-  addGlobalSidebarContentPanel(
-    buttonDescriptor: Kefir.Observable<Object>
-  ): Promise<?ContentPanelViewDriver>,
-  getButterBarDriver(): any,
-  setButterBar(butterBar: any): void,
-  isRunningInPageContext(): boolean,
-  showAppIdWarning(): void,
-  createModalViewDriver(options: Object): Object,
-  createMoleViewDriver(options: Object): MoleViewDriver,
-  createTopMessageBarDriver(options: Object): Object,
-  createDrawerViewDriver(options: DrawerViewOptions): DrawerViewDriver,
-  createBackdrop(zIndex?: number, target?: HTMLElement): Backdrop,
-  getStopper(): Kefir.Observable<any>,
-  destroy(): void,
-
-  getSelectedThreadRowViewDrivers(): $ReadOnlyArray<ThreadRowViewDriver>,
-  registerThreadRowViewSelectionHandler(handler: () => any): () => void,
-};
+export type Driver = GmailDriver;
 
 export type ButterBarDriver = {
   getNoticeAvailableStream(): Kefir.Observable<any, any>,
